Move errors to namespaces
