+++ lib/npmdc/checkers/yarn/yarn_installation_checker.rb
@@ -1,42 +1,47 @@
+require 'npmdc/checkers/yarn/errors'
+
 module Npmdc
   module Checkers
-    class YarnInstallationChecker
-      attr_reader :yarn_command
+    module Yarn
+      class YarnInstallationChecker
+        include Errors
+        attr_reader :yarn_command
 
-      def initialize(yarn_command)
-        @yarn_command = yarn_command
-      end
-
-      def check_yarn_is_installed
-        if yarn_command == 'yarn'
-          check_yarn_is_installed_default
-        else
-          check_yarn_is_installed_on_the_given_path
+        def initialize(yarn_command)
+          @yarn_command = yarn_command
         end
-      end
 
-      private
-
-      def os_extensions
-        ENV['PATHEXT'] ? ENV['PATHEXT'].split(';') : ['']
-      end
-
-      def check_yarn_is_installed_on_the_given_path
-        os_extensions.each do |ext|
-          yarn = "#{yarn_command}#{ext}"
-          return true if File.executable?(yarn) && !File.directory?(yarn)
-        end
-        raise(Npmdc::Errors::YarnNotInstalledError, yarn_command: yarn_command)
-      end
-
-      def check_yarn_is_installed_default
-        ENV['PATH'].split(File::PATH_SEPARATOR).each do |path|
-          os_extensions.each do |ext|
-            yarn = File.join(path, "#{yarn_command}#{ext}")
-            return true if File.executable?(yarn) && !File.directory?(yarn)
+        def check_yarn_is_installed
+          if yarn_command == 'yarn'
+            check_yarn_is_installed_default
+          else
+            check_yarn_is_installed_on_the_given_path
           end
         end
-        raise(Npmdc::Errors::YarnNotInstalledError, yarn_command: yarn_command)
+
+        private
+
+        def os_extensions
+          ENV['PATHEXT'] ? ENV['PATHEXT'].split(';') : ['']
+        end
+
+        def check_yarn_is_installed_on_the_given_path
+          os_extensions.each do |ext|
+            yarn = "#{yarn_command}#{ext}"
+            return true if File.executable?(yarn) && !File.directory?(yarn)
+          end
+          raise(YarnNotInstalledError, yarn_command: yarn_command)
+        end
+
+        def check_yarn_is_installed_default
+          ENV['PATH'].split(File::PATH_SEPARATOR).each do |path|
+            os_extensions.each do |ext|
+              yarn = File.join(path, "#{yarn_command}#{ext}")
+              return true if File.executable?(yarn) && !File.directory?(yarn)
+            end
+          end
+          raise(YarnNotInstalledError, yarn_command: yarn_command)
+        end
       end
     end
   end
